Improvements

+++ Core/Src/Subsystems/event_center.cpp
@@ -21,7 +21,8 @@ void EventCenter::init() {
 
 UARM_Events_t EventCenter::wait_events(UARM_Events_t wait_events,
                                        uint32_t timeout) {
-    return xEventGroupWaitBits(events_group, wait_events, wait_events, pdTRUE,
+    return xEventGroupWaitBits(events_group, static_cast<uint32_t>(wait_events),
+                               static_cast<uint32_t>(wait_events), pdTRUE,
                                timeout);
 }
